# internal/docs/swagger.yaml
definitions:
...
  api.CreateProjectRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ProjectResponse:
    properties:
      asset_count:
        type: integer
      color:
        type: string
      cover_asset_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
...
  api.UpdateProjectRequest:
    properties:
      color:
        type: string
      cover_asset_id:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  api.ValidationErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/projects:
    get:
      description: Returns all projects in the workspace, each with an <code>asset_count</code>
        field reflecting the number of assets currently assigned to it.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ProjectResponse'
            type: array
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Creates a new project container. Projects group assets and folders
        together. An optional color (hex) and description can be provided at creation
        time.
      parameters:
      - description: Project details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.ProjectResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - Projects
  /api/v1/projects/{id}:
    delete:
      description: Permanently deletes the project. Assets that belonged to the project
        are <em>not</em> deleted — their <code>project_id</code> is set to null so
        they remain accessible in the unfiltered asset library.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - Projects
    get:
      description: Returns the project with the given ID. Returns 404 if the project
        does not exist or belongs to a different workspace.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ProjectResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates the project's name, description, color, or cover asset.
        All fields are optional — only the fields present in the request body are
        updated; omitted fields retain their current values.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ProjectResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects
...
swagger: "2.0"